fix travis build: use phpenv version in php-fpm include path (#21)
For some PHP versions, the version used in the php-fpm.conf.default
include path doesn't match the output of 'phpenv version-name'.

For example, as of this writing, with the 7.1 version, the config file
references ~/.phpenv/versions/7.1.11, but 'phpenv version-name' returns
"7.1", which means the php-fpm service does not reference the config
files that are copied in during the build setup.

Now we write our own php-fpm.conf which uses the path from
'phpenv version-name', which guarantees our config files are used.

In addition, we purge the 'www.conf' file which is provided by phpenv
and interferes with our configurations (duplicate listening port,
unknown group, etc.).