Inject socket ID to socket via constructor
Let the socket collection create a new socket ID and instantiate the
socket class with it instead of creating it internally on creation.

This is a preparation to replace the socket ID as the actual request ID
for the fastCGI packages with a dedicated request ID for each request.

This change is needed to allow multiplexing requests.
(Multiple requests on one stream)