Make the AbstractRequest content immutable
This commit removes the setContent setter from the AbstractRequest class.
The only way to set the content of a request is now through the constructor.
Once the content for a request is defined, it cannot be changed.
If new content has to be sent, a new request should be made.