Remove the multi-request stream_select call from Client
This method had a bug where, as long as no error was thrown, or timeout
reached, it would report that all sockets inside a collection had responses
even if some of them did not. Checking each socket individually in a loop
is a performance hit, but the only way to report accuratly if a socket
has a response ready.